Updating to use cocos2d 3.4.9
NB: There are a lot of linker warnings thrown due to a bad configuration in the ObjectAL submodule of cocos2d. It's easy to fix, but without write access to that repo it is hard to make a shareable project that is fixed, so for now I've left the module as is, with warnings and all.

NB2: Updated minimum iOS level to 8 as I don't have any earlier hardware to test on and XCode 7 doesn't support iOS7 simulators.